feat: Add album music list and related widgets
- Implemented AlbumMusicListMusic widget to display music list with loading states and search functionality.
- Created AlbumMusicMainAppbar for the main app bar in the album music section.
- Added AlbumMusicSearchBar for searching music within the album.
- Introduced AlbumMusicShuffleAddAll for shuffling and adding all music to the playlist.
- Developed AlbumMusicTitleArtist to show album title and artist information.
- Refactored album music list modes to use new widget structure.
- Removed obsolete AzlistViewMode.
- Added AlbumMusicDeleteDialog for confirming music deletion.
- Created AlbumMusicEffectTapModal for tap effects on music list items.
- Implemented AlbumMusicIndexNumberList for displaying download progress.
- Developed AlbumMusicModal for music options (play, add to playlist, download, delete).
- Updated SearchAlbumScreen to handle keyboard focus properly on back button press.
