+++ lib/features/album_music/widgets/album_music_list_mode/azlist_mode/album_music_azlist_list.dart
@@ -5,8 +5,8 @@ import 'package:cybeat_music_player/core/controllers/audio_state_controller.dart
 import 'package:cybeat_music_player/common/utils/capitalize.dart';
 import 'package:cybeat_music_player/core/controllers/music_player_controller.dart';
 import 'package:cybeat_music_player/core/models/music.dart';
-import 'package:cybeat_music_player/features/album_music/widgets/album_music_index_number_list.dart';
-import 'package:cybeat_music_player/features/album_music/widgets/album_music_modal.dart';
+import 'package:cybeat_music_player/features/album_music/widgets/album_music_widget_musiclist/album_music_index_number_list.dart';
+import 'package:cybeat_music_player/features/album_music/widgets/album_music_widget_musiclist/album_music_modal.dart';
 import 'package:cybeat_music_player/common/widgets/spectrum_animation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
