+++ lib/features/album_music/widgets/album_music_widget_musiclist/album_music_modal.dart
@@ -3,8 +3,8 @@ import 'package:cybeat_music_player/common/utils/toast.dart';
 import 'package:cybeat_music_player/core/controllers/music_download_controller.dart';
 import 'package:cybeat_music_player/core/controllers/audio_state_controller.dart';
 import 'package:cybeat_music_player/core/controllers/music_player_controller.dart';
-import 'package:cybeat_music_player/features/album_music/widgets/album_music_delete_dialog.dart';
-import 'package:cybeat_music_player/features/album_music/widgets/album_music_effect_tap_modal.dart';
+import 'package:cybeat_music_player/features/album_music/widgets/album_music_widget_musiclist/album_music_delete_dialog.dart';
+import 'package:cybeat_music_player/features/album_music/widgets/album_music_widget_musiclist/album_music_effect_tap_modal.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:hexcolor/hexcolor.dart';
